Move mRNA tab state into a PatientViewPlotsStore sub-store
Extract the plot-specific remoteData/observables out of PatientViewPageStore into a new PatientViewPlotsStore. The page store instantiates and holds it; the sub-store references the parent for shared context (studyId, molecular profiles). The parent type is used only as a constructor-param annotation, so there is no runtime import cycle.
